addons/mirror-gesture/requirements.txt
opencv-python-headless==4.10.0.84
mediapipe==0.10.20
paho-mqtt==2.1.0
